Update Underware_Item_Holder_Clamshell_Style.scad
Hey Andy, thanks for your work on the Underware project.

I tweaked this clamshell scad to add the Global_Color, like you have in other scripts. I tested it in MakerWorld with many different options and it seems to color the shapes properly.

Also, I found a small bug in this version where the cutouts of the clamshell were being removed from the opposite side of the top shape, rather than being mirrored. In a previous version you had zflip() in place of rot([0,180,0]) on line 248 and that seems to have solved the issue. 

Thanks,
~Ryan