Adds an option to align transition channel to edges
Before this PR if a channel was smaller than the other end, it would be centered and then shifted half a unit if needed to align it to the grid.

I had a need to align it to the edge instead of the center so this PR adds a option to align it left or right edge instead of attempting to center it.